server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "AI Powered Hardware Store Management System"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "httpx>=0.28.1",
    "imagekitio>=5.1.2",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
